Adding Symbol me
Convert all your similar layers to symbols with a single click.

This plug-in will convert the select layer to a symbol and updates all the other layers with same name with symbol just created.

This is really helpful when you want to covert a bunch of scattered (nested deep down and its painful to select all of them one by one) layers to symbols